Separated logic for temperature from model to view entity. Grouped weather specific models to weather package in common util module.
