+++ common_util/src/main/java/com/zywczas/commonutil/weather/TemperatureColor.kt
@@ -1,0 +1,9 @@
+package com.zywczas.commonutil.weather
+
+import androidx.compose.ui.graphics.Color
+
+enum class TemperatureColor(val value: Color) {
+    Cold(Color(0xFF00528D)),
+    Neutral(Color(0xFF000000)),
+    Hot(Color(0xFFD01717));
+}
