+++ feature_forecast_place/src/main/java/com/zywczas/featureforecastplace/viewmodel/PlaceForecastViewEntity.kt
@@ -2,8 +2,8 @@ package com.zywczas.featureforecastplace.viewmodel
 
 import androidx.compose.ui.graphics.Color
 import com.zywczas.commoncompose.components.KeyValueViewEntity
-import com.zywczas.commoncompose.theme.TemperatureColor
-import com.zywczas.commonutil.WeatherCondition
+import com.zywczas.commonutil.weather.TemperatureColor
+import com.zywczas.commonutil.weather.WeatherCondition
 
 internal data class PlaceForecastViewEntity(
     val toolbarTitle: String = "",
